#620: Refactor tooling plugin
