+++ docs/changelog/2018.md
@@ -1,5 +1,7 @@
-v3.0.0-beta.36 - [January 15, 2018](https://github.com/lando/lando/releases/tag/v3.0.0-beta.36)
------------------------------------
+v3.0.0-beta.36 - In Development
+-------------------------------
+
+**THIS RELEASE CONTAINS SIGNIFICANT CODE REFACTORING TO HELP WITH TESTING. YOU MAY WANT TO HOLD OFF FOR A FEW VERSIONS IF YOU WANT TO BE SUPER SAFE ABOUT THINGS!**
 
 **SERIOUSLY, READ THE DOCS!: https://docs.devwithlando.io/**
 
@@ -7,7 +9,9 @@ v3.0.0-beta.36 - [January 15, 2018](https://github.com/lando/lando/releases/tag/
 * Documented common .lando.yml syntax more explicitly [#643](https://github.com/lando/lando/issues/643)
 * Finally moved over to all the `eslint` [#620](https://github.com/lando/lando/issues/620)
 * Fixed bug where offline mode was failing on metrics check [#630](https://github.com/lando/lando/issues/630)
+* Fixed bug where `php 5.3 apache` was not starting correctly [#652](https://github.com/lando/lando/issues/652)
 * Provided some corrective action around `proxy` start failures [#632](https://github.com/lando/lando/issues/632)
+* Refactored the code to increase testibility and reduce complexity and dependents [#620](https://github.com/lando/lando/issues/620)
 
 v3.0.0-beta.35 - [January 8, 2018](https://github.com/lando/lando/releases/tag/v3.0.0-beta.35)
 ----------------------------------
