+++ lib/config.js
@@ -52,7 +52,7 @@ var getSysConfRoot = function() {
 exports.merge = function(old, fresh) {
   return _.mergeWith(old, fresh, function(s, f) {
     if (_.isArray(s)) {
-      return s.concat(f);
+      return _.uniq(s.concat(f));
     }
   });
 };
