+++ plugins/lando-tooling/index.js
@@ -8,10 +8,61 @@
 
 module.exports = function(lando) {
 
-  // Add tooling settings to the global config
-  require('./lib/bootstrap')(lando);
+  // Modules
+  var _ = lando.node._;
 
-  // The tooling service
-  require('./lib/tooling')(lando);
+  // Add tooling module to lando
+  lando.events.on('post-bootstrap', 2, function(lando) {
+
+    // Log
+    lando.log.info('Initializing tooling plugin');
+
+    // Add the SSH command
+    lando.tasks.add('ssh', require('./tasks/ssh')(lando));
+
+    // Add services to lando
+    lando.tooling = require('./tooling')(lando);
+
+  });
+
+  // Try to detect additional commands if we have app context
+  lando.events.on('post-bootstrap', function(lando) {
+
+    // Try to determine app context so we can load in any tooling commands that
+    // are defined there
+    return lando.app.get()
+
+    // If we have an app with a tooling section let's do this
+    .then(function(app) {
+      if (app && app.config.tooling && !_.isEmpty(app.config.tooling)) {
+
+        // Log
+        lando.log.verbose('Additional tooling detected for app %s', app.name);
+
+        // Loop through each tool
+        _.forEach(app.config.tooling, function(task, name) {
+          if (_.isObject(task)) {
+
+            // Log
+            lando.log.verbose('Adding app cli task %s', name);
+
+            // Build our config
+            var config = task;
+            task.app = app;
+            task.name = name;
+
+            // Build and add the task
+            lando.tasks.add(name, lando.tooling.build(config));
+
+          }
+        });
+
+        // Log.
+        lando.log.verbose('App tooling loaded.');
+
+      }
+    });
+
+  });
 
 };
