+++ plugins/lando-tooling/tooling.js
@@ -13,34 +13,7 @@ module.exports = function(lando) {
   var path = require('path');
   var Promise = lando.Promise;
   var format = require('util').format;
-
-  /*
-   * Helper to process args
-   */
-  var largs = function(config) {
-
-    // We assume pass through commands so let's use argv directly and strip out
-    // the first three assuming they are [node, lando.js, options.name]
-    var argopts = process.argv.slice(3);
-
-    // Arrayify the command if needed
-    // @todo: this could probably be improved even more to handle chained commands
-    if (_.has(config, 'cmd') && typeof config.cmd === 'string') {
-      config.cmd = config.cmd.split(' ');
-    }
-
-    // Shift on our command
-    argopts.unshift(config.cmd || config.name);
-
-    // Check to see if we have global lando opts and remove them if we do
-    if (_.indexOf(argopts, '--') > 0) {
-      argopts = _.slice(argopts, 0, _.indexOf(argopts, '--'));
-    }
-
-    // Return
-    return _.flatten(argopts);
-
-  };
+  var utils = require('./lib/utils');
 
   /*
    * The task builder
@@ -84,12 +57,12 @@ module.exports = function(lando) {
       .then(function(exists) {
 
         // Helper
-        var decartes = function(current, exist) {
+        var descartes = function(current, exist) {
           return current && exist;
         };
 
         // Pass to should start if something doesnt exist
-        if (!_.reduce(exists, decartes, true)) {
+        if (!_.reduce(exists, descartes, true)) {
           return true;
         }
 
@@ -97,7 +70,7 @@ module.exports = function(lando) {
         else {
           return Promise.all(runCheck)
           .then(function(isRunning) {
-            return !_.reduce(isRunning, decartes, true);
+            return !_.reduce(isRunning, descartes, true);
           });
         }
 
@@ -115,7 +88,7 @@ module.exports = function(lando) {
       .then(function() {
 
         // Build the command
-        var cmd = largs(config);
+        var cmd = utils.largs(config);
 
         // Break up our app root and cwd so we can get a diff
         var appRoot = config.app.root.split(path.sep);
